Rename agama-dbus-server "locale" to "l10n"
