+++ rust/agama-dbus-server/src/l10n.rs
